Update with insertions testet on my desktop
Adding to the header a bottom in two different sized columns.
The column on the left is smaller. It contains a picture with a map which shows the location of our event.
The column on the right is the bigger one containing a box enabling one to leave text. The button clear enables one to erase all text. The text won't be shown after reloading the page.